pyproject.toml
[project]
name = "dataseed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Juan Sebastian Peña", email = "jspenaq@unal.edu.co" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "fastapi>=0.116.1",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "coverage>=7.10.3",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.8",
]

# Ruff configuration

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "Q",  # flake8-quotes
    "S",  # flake8-bandit
    "B",  # flake8-bugbear
    "UP",  # pyupgrade
    "TC",  # flake8-type-checking
    "TD",  # flake8-todos
    "C4",  # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "ANN",  # flake8-annotations
    "SIM",  # flake8-simplify
    "RET",  # flake8-return
    "COM",  # flake8-commas
    "ICN",  # flake8-import-conventions
    "TID",  # flake8-tidy-imports
    "PTH",  # flake8-use-pathlib
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN001", "ANN201", "ANN202", "E402", "E501", "E711"]


# Coverage configuration

[tool.coverage.run]
branch = true
source = [
    "app",  # Include the app package
]

[tool.coverage.report]
# Regexes for lines to exclude from consideration
omit = [
    "tests/*",  # Exclude all test files
    "*/__init__.py",  # Exclude __init__.py files
]

show_missing = true

[tool.coverage.html]
directory = "coverage_html_report"

